Update code for changed data type

+++ SteamCloudFileManager/IRemoteStorage.cs
@@ -7,7 +7,7 @@ namespace SteamCloudFileManager
     {
         IRemoteFile GetFile(string name);
         List<IRemoteFile> GetFiles();
-        bool GetQuota(out int totalBytes, out int availableBytes);
+        bool GetQuota(out ulong totalBytes, out ulong availableBytes);
         bool IsCloudEnabledForAccount { get; }
         bool IsCloudEnabledForApp { get; set; }
     }
