+++ SteamCloudFileManager/RemoteStorage.cs
@@ -80,7 +80,7 @@ namespace SteamCloudFileManager
             return new RemoteFile(this, name.ToLowerInvariant());
         }
 
-        public bool GetQuota(out int totalBytes, out int availableBytes)
+        public bool GetQuota(out ulong totalBytes, out ulong availableBytes)
         {
             checkDisposed();
             return SteamRemoteStorage.GetQuota(out totalBytes, out availableBytes);
